feat: add permission gift product

+++ Modules/Gift/Http/Controllers/GiftProductController.php
@@ -3,7 +3,9 @@
 namespace Modules\Gift\Http\Controllers;
 
 use App\Enums\TargetType;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Support\Renderable;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Modules\Gift\Repositories\GiftProductRepository;
@@ -12,6 +14,8 @@ use Modules\Product\Repositories\VariationRepository;
 
 class GiftProductController extends Controller
 {
+    use AuthorizesRequests;
+
     /** @var \Modules\Gift\Repositories\GiftProductRepository */
     protected $giftProductRepository;
 
@@ -39,10 +43,18 @@ class GiftProductController extends Controller
      */
     public function index(Request $request, $gift_id)
     {
-        $search = $request->input('search');
-        $gift_products = $this->giftProductRepository->paginateByGiftId($gift_id, $search);
+        try {
+            $this->authorize('gift_product:browse');
 
-        return view('gift::product.index', compact('gift_products', 'gift_id'));
+            $search = $request->input('search');
+            $gift_products = $this->giftProductRepository->paginateByGiftId($gift_id, $search);
+    
+            return view('gift::product.index', compact('gift_products', 'gift_id'));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.index')
+            ->with('danger', __('notification.permission.fail', ['action' => 'browse gift product']));
+        }
     }
 
     /**
@@ -51,17 +63,25 @@ class GiftProductController extends Controller
      */
     public function create($gift_id)
     {
-        $form = [
-            'title'     => 'Create',
-            'url'       => route('admin.gift.product.store', $gift_id),
-            'method'    => 'POST',
-        ];
+        try {
+            $this->authorize('gift_product:add');
 
-        $products = $this->productRepository->all();
-        $variants = $this->variantRepository->all();
-        $target_types = TargetType::getObject();
-
-        return view('gift::product.create', compact('form', 'products', 'variants', 'target_types', 'gift_id'));
+            $form = [
+                'title'     => 'Create',
+                'url'       => route('admin.gift.product.store', $gift_id),
+                'method'    => 'POST',
+            ];
+    
+            $products = $this->productRepository->all();
+            $variants = $this->variantRepository->all();
+            $target_types = TargetType::getObject();
+    
+            return view('gift::product.create', compact('form', 'products', 'variants', 'target_types', 'gift_id'));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'add gift product']));
+        }
     }
 
     /**
@@ -71,16 +91,24 @@ class GiftProductController extends Controller
      */
     public function store(Request $request, $gift_id)
     {
-        $request->validate([
-            'target_id'     => 'required',
-            'target_type'   => 'required',
-        ]);
+        try {
+            $this->authorize('gift_product:add');
 
-        $this->giftProductRepository->create($request->all(), ['gift_id' => $gift_id]);
-
-        return redirect()
-        ->route('admin.gift.product.index', $gift_id)
-        ->with('success', __('notification.create.success', ['model' => 'gift product']));
+            $request->validate([
+                'target_id'     => 'required',
+                'target_type'   => 'required',
+            ]);
+    
+            $this->giftProductRepository->create($request->all(), ['gift_id' => $gift_id]);
+    
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('success', __('notification.create.success', ['model' => 'gift product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'add gift product']));
+        }
     }
 
     /**
@@ -90,18 +118,26 @@ class GiftProductController extends Controller
      */
     public function edit($gift_id, $id)
     {
-        $form = [
-            'title'     => 'Edit',
-            'url'       => route('admin.gift.product.update', ['gift_id' => $gift_id, 'id' => $id]),
-            'method'    => 'PUT',
-        ];
+        try {
+            $this->authorize('gift_product:edit');
 
-        $products = $this->productRepository->all();
-        $variants = $this->variantRepository->all();
-        $target_types = TargetType::getObject();
-        $gift_product = $this->giftProductRepository->find($id);
-
-        return view('gift::product.edit', compact('form', 'products', 'variants', 'target_types', 'gift_id', 'gift_product'));
+            $form = [
+                'title'     => 'Edit',
+                'url'       => route('admin.gift.product.update', ['gift_id' => $gift_id, 'id' => $id]),
+                'method'    => 'PUT',
+            ];
+    
+            $products = $this->productRepository->all();
+            $variants = $this->variantRepository->all();
+            $target_types = TargetType::getObject();
+            $gift_product = $this->giftProductRepository->find($id);
+    
+            return view('gift::product.edit', compact('form', 'products', 'variants', 'target_types', 'gift_id', 'gift_product'));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'edit gift product']));
+        }
     }
 
     /**
@@ -112,16 +148,24 @@ class GiftProductController extends Controller
      */
     public function update(Request $request, $gift_id, $id)
     {
-        $request->validate([
-            'target_id'     => 'required',
-            'target_type'   => 'required',
-        ]);
+        try {
+            $this->authorize('gift_product:edit');
 
-        $this->giftProductRepository->update($id, $request->all());
-
-        return redirect()
-        ->route('admin.gift.product.index', $gift_id)
-        ->with('success', __('notification.update.success', ['model' => 'gift product']));
+            $request->validate([
+                'target_id'     => 'required',
+                'target_type'   => 'required',
+            ]);
+    
+            $this->giftProductRepository->update($id, $request->all());
+    
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('success', __('notification.update.success', ['model' => 'gift product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'edit gift product']));
+        }
     }
 
     /**
@@ -131,10 +175,18 @@ class GiftProductController extends Controller
      */
     public function destroy($gift_id, $id)
     {
-        $this->giftProductRepository->delete($id);
+        try {
+            $this->authorize('gift_product:delete');
 
-        return redirect()
-        ->route('admin.gift.product.index', $gift_id)
-        ->with('success', __('notification.delete.success', ['model' => 'gift product']));
+            $this->giftProductRepository->delete($id);
+    
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('success', __('notification.delete.success', ['model' => 'gift product']));
+        } catch (AuthorizationException $exception) {
+            return redirect()
+            ->route('admin.gift.product.index', $gift_id)
+            ->with('danger', __('notification.permission.fail', ['action' => 'delete gift product']));
+        }
     }
 }
